+++ src/SCQAT/Language/PHP/Analyzer/PhpCs.php
@@ -23,14 +23,20 @@ class PhpCs extends \SCQAT\AnalyzerAbstract
     {
         $result = new \SCQAT\Result();
 
-        $processBuilder = new ProcessBuilder(array("php", $context->vendorDirectory."bin/phpcs", "--standard=PSR2", "-n", $analyzedFileName));
+        $processBuilder = new ProcessBuilder(array("php", $context->vendorDirectory."bin/phpcs", "--standard=PSR2", "--report=xml", "-n", $analyzedFileName));
         $process = $processBuilder->getProcess();
         $process->run();
 
         if (!$process->isSuccessful()) {
             $result->isSuccess = false;
             $result->value = "KO";
-            $result->description = trim($process->getOutput());
+            $description = "";
+            $simpleXml = new \SimpleXMLElement(trim($process->getOutput()));
+            foreach ($simpleXml->file[0] as $error) {
+                // Building SCQAT file error messages (one per line)
+                $description .= ($description != "" ? "\n" : "").":".$error["line"]."  (error) ".$error;
+            }
+            $result->description = trim($description);
         } else {
             $result->isSuccess = true;
             $result->value = "OK";